fix(collaboration): release the materializer when the bootstrap tail throws
The factory's catch destroyed only the registry, so a throw from
materializing, serializing, or session construction leaked the
materializer's observers on the caller's document. The tail now hands the
materializer back on the way out and serializes before the session exists.
Undo and redo re-check the gate after their flush, because the flush can
refuse the queued journal and take the session to terminal error. The
teardown census in tests now counts the blob map, the seed-record array,
and doc-level afterTransaction handlers. The identity validation and
awareness codec move to document-awareness.ts to stay under the line cap.
